requirements.txt
allure-pytest==2.13.5
Faker==28.0.0
pydantic==2.8.2
pytest==8.3.2
python-dotenv==1.0.1
requests==2.32.3
